[BEAM-951] FileBasedSink: merge FileOperations into IOChannelFactory.
