+++ sdks/java/core/src/main/java/org/apache/beam/sdk/util/IOChannelFactory.java
@@ -23,6 +23,7 @@ import java.nio.channels.ReadableByteChannel;
 import java.nio.channels.WritableByteChannel;
 import java.nio.file.Path;
 import java.util.Collection;
+import java.util.List;
 
 /**
  * Defines a factory for working with read and write channels.
@@ -31,7 +32,10 @@ import java.util.Collection;
  *
  * <p>See <a href="http://docs.oracle.com/javase/7/docs/api/java/nio/channels/package-summary.html"
  * >Java NIO Channels</a>
+ *
+ * @deprecated This is under redesign, see: https://issues.apache.org/jira/browse/BEAM-59.
  */
+@Deprecated
 public interface IOChannelFactory {
 
   /**
@@ -103,4 +107,22 @@ public interface IOChannelFactory {
 
   /** Converts the given string to a {@link Path}. */
   Path toPath(String path);
+
+  /**
+   * Copies a collection of files from one location to another.
+   *
+   * <p>The number of source filenames must equal the number of destination filenames.
+   *
+   * @param srcFilenames the source filenames.
+   * @param destFilenames the destination filenames.
+   */
+  void copy(List<String> srcFilenames, List<String> destFilenames) throws IOException;
+
+  /**
+   * Removes a collection of files or directories.
+   *
+   * <p>Directories are required to be empty. Non-empty directories will not be deleted,
+   * and this method may return silently or throw an exception.
+   */
+  void remove(Collection<String> filesOrDirs) throws IOException;
 }
